+++ src/Resources/ActiveCampaignFieldOptionsResource.php
@@ -62,7 +62,7 @@ class ActiveCampaignFieldOptionsResource extends ActiveCampaignResource
      */
     protected function responseCast(array $response): array
     {
-        $responseCast = $response['fieldOption'];
+        $responseCast = $response['fieldOption'] ?? [];
 
         unset($responseCast['links']);